requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
clr-loader==0.2.5
colorzero==2.0
common==0.1.2
Django==4.2.1
frozenlist==1.3.3
gpiozero==1.6.2
idna==3.4
image==1.5.33
magic-filter==1.0.9
markdown-it-py==2.2.0
mdurl==0.1.2
mouse==0.7.1
multidict==6.0.4
Pillow==9.5.0
psutil==5.9.5
py-cpuinfo==9.0.0
pycparser==2.21
Pygments==2.15.1
pytz==2023.3
pywin32==306
requests==2.30.0
rich==13.3.5
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
yarl==1.9.2
